% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.r
\name{assert-is}
\alias{is.error}
\alias{is.time}
\alias{is.date}
\title{Missing is functions.}
\usage{
is.error(x)

is.time(x)

is.date(x)
}
\arguments{
\item{x}{object to test}
}
\description{
Missing is functions.
}
\examples{
a <- Sys.time()
is.time(a)
b <- Sys.Date()
is.date(b)
c <- try(stop("!!"))
is.error(c)
}
\seealso{
Other assertions: \code{\link{are_equal}},
  \code{\link{is.scalar}}, \code{\link{noNA}},
  \code{\link{not_empty}}
}
