% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert-packages.R
\name{assert_packages}
\alias{assert_packages}
\title{Assert that packages are loaded.}
\usage{
assert_packages(packages, ...)
}
\arguments{
\item{packages}{character vector of package names to check}

\item{...}{character package names to check}
}
\value{
`logical` indicating success
}
\description{
This function can be useful in writing functions that rely on packages
that you "suggest". It asserts that these packages are available, and 
throws an informative error for those packages that are not.
}
\examples{
  # put package names in a character vector
  assert_packages(c("base", "utils"))

  # or by themselves
  assert_packages("base", "utils")

\dontrun{
  # intentionally invokes error-behavior
  assert_packages("utils2")
}
}
\seealso{
\href{http://r-pkgs.had.co.nz/description.html#dependencies}{R Packages book}
}
