% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is-prime.R
\name{is.prime}
\alias{is.prime}
\title{Test if a scalar is prime}
\usage{
is.prime(x)
}
\arguments{
\item{x}{The scalar to test}
}
\value{
TRUE when \code{x} is a prime, FALSE when \code{x} is not a prime.
}
\description{
If the algorithm finds a prime \code{a} for which
   \code{x \%\% a == 0}, then \code{a} is not a prime. If \code{x} is not a
   prime then there exist two primes \code{a} and \code{b} for which
   \code{x <- a * b}. Assuming \code{a <= b}, then \code{a <= sqrt(x)}. Hence
   the algorithm considers only those \code{a <= sqrt(x)}. It reliase on
   precompute primes to make it as fast as possible.
}
\author{
Thierry Onkelinx
}

