% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assertions.r
\name{has_attr}
\alias{\%has_attr\%}
\alias{\%has_name\%}
\alias{has_attr}
\alias{has_name}
\title{Has attribute or name?}
\usage{
has_attr(x, which)

x \%has_attr\% which

has_name(x, which)

x \%has_name\% which
}
\arguments{
\item{x}{object to test}

\item{which}{name or attribute}
}
\description{
Has attribute or name?
}
\examples{
has_attr(has_attr, "fail")
x <- 10
x \%has_attr\% "a"

y <- list(a = 1, b = 2)
see_if(y \%has_name\% "c")
}

