% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.r
\name{not_empty}
\alias{not_empty}
\title{Check an object doesn't have any empty dimensions}
\usage{
not_empty(x)
}
\arguments{
\item{x}{object to test}
}
\description{
Check an object doesn't have any empty dimensions
}
\examples{
not_empty(numeric())
not_empty(mtcars[0, ])
not_empty(mtcars[, 0])
}
\seealso{
Other assertions: \code{\link{are_equal}},
  \code{\link{is.error}}, \code{\link{is.scalar}},
  \code{\link{noNA}}
}
